Finish unifiying x8632 FFI stuff.
Most operating systems use the vanilla i386 ABI conventions.

Darwin uses a special rule for returning small structures. Call the
CALL-STRUCT-RETURN-BY-IMPLICIT-ARG-FUNCTION (from the FTD) in
EXPAND-FF-CALL and GENERATE-CALLBACK-BINDINGS in order to tell when we
need to deal with Darwin struct return.

In GENERATE-CALLBACK-BINDINGS, return an 8th value (number of bytes of
outgoing arguments pushed) for Windows's benefit;  other backends
should just ignore it.