The lisp assembler generates slightly different code than GAS does for x8632 instructions that have a memory operand that contains only a 32-bit displacement (but no scale factor or index/base registers.) pc_luser_xp() expects the shorter encodings for those affected instructions that're used in the consing sequence, and may therefore adjust the PC of a thread that's interrupted during consing by too few bytes if a longer-than-expected instruction is used.
If LISP_ASSEMBLER_EXTRA_SIB_BYTE is defined, try to detect cases where
this extra byte is present and account for it when adjusting the PC
of any thread interrupted in the middle of allocation.

LISP_ASSEMBLER_EXTRA_SIB_BYTE is defined on x8632; if/when we fix
the assembler and bootstrap that change, we can stop defining it
and remove the conditionalized code.