Functions that parse namestrings have called GET-SSTRING on their string arguments so that they're sure that they're operating on a bounded SIMPLE-STRING.  Those parsing functions (hopefully all of them) now call GET-PATHNAME-SSTRING instead. On non-Windows platforms, this just calls GET-SSTRING and returns its results; on Windows, it checks to see if the bounded simple-string contains backslashes and if so, it returns a copy with the backslashes replaced with forward-slashes.
This is one way to allow pathname functions to handle Windows pathnames
that use #\\ as a directory separator.  (The other way would be to 
change a lot of ancient namestring-parsing code to handle either #\\
or #\/ as a directory separator, and this approach is certainly simpler.)
The extra consing should only happen if user code does something like:

(let* ((home-string (getenv "HOME")))
  (if home-string
    (pathname-name home-string)))

System code that deals with namestrings returned by the OS can generally
destructively change any backslashes in that namestring to forward slashes
before passing the string to pathname-parsing functions.