Try to arrange it so that calling structure slot readers doesn't cause warnings about unknown types in the slot :type.  There are times where this might indicate a missed optimization opportunity, but most of the time it's just a gratuitous warning.
Also:

make defstruct check for validity (but not definedness) of the slot type.

Fix the undefined type warnings to record the whole type, not just the thing that's undefined,
in case there are more than one undefined references in the type.

Make proclaim again signal an error (not just warn) if the declaration is invalid.