When handling a write to a watched object, try to be more careful about assuming that the object is in a static gc area.  This is the idea behind using offsets from the tagged object pointer, rather than an address passed as an integer.
In lieu of using the passed in address to locate the watched object,
assume that the kernel will have pushed the watched object onto the lisp
stack immediately before the xcf.

Don't offer to emulate writes on Windows platforms, since they don't
have #_mprotect.  (This is just laziness.  Windows has #_VirtualProtect, so
it's just a matter of using it correctly and testing it.)