Added a new function GET-ALLOCATION-SENTINEL and a new :START arg to HEAP-UTILIZATION.  get-allocation-sentinel returns an object which can be passed as the :start arg to heap-utilization, in which case heap-utilization will only scan objects allocated after the call to get-allocation-sentinel.  I.e.
   (let ((sentinel (get-allocation-sentinel)))
     (do-things)
     (heap-utilization :start sentinel))

will show a breakdown of still-reachable objects allocated by (do-things).

Note that both get-allocation-sentinel and heap-utilization do a full
gc by default.  They both take a gc-first arg that can be used to turn
that off.  This will result in some fuzziness in the results, no larger
than the size of the youngest ephemeral generation, so is only useful
when the overall allocation is significantly larger than that.