In %bignum-random, call cl:random to generate a full digit's worth of random bits.  Neither the new generator nor the old one generates a full 32 random bits.  (We do cheat a little with the current generator, but it's not clear that our cheat is justifiable.)
It may be best to do bignums 16 bits at a time, just as we do on 32-bit
ports, at least until we put in a random number generator that
produces longer values.

One possible issue is that higher quality generators are more expensive
than the dead simple generator that we've historically used.  It may be
worth a little bit shuffling effort to use all of the random bits we
get from the generator rather than throwing all but 16 of them out.