Improve CL:RANDOM.
The new generator is the MRG321k3p generator described in
P. L'Ecuyer and R. Touzin, "Fast Combined Multiple Recursive
Generators with Multipliers of the form a = +/- 2^d +/- 2^e"",
Proceedings of the 2000 Winter Simulation Conference, Dec. 2000,
683--689.

It has a period of about 2^185 and produces output of much higher
statistical quality than the previous generator.

Performance of the new generator should generally be comparable to that
of the old generator, despite the fact that the new generator does
a lot more work.