Yet another scheme for subprim calls.  Go ahead and laugh. Details: - ba/bla are new pseudo instructions, identical to b/bl except that   their operands are subprim names (or addresses). - for each subprim name/address referenced in a ba/bla instruction,   the assembler generates an:
  (ldr pc (:= data-word-containing-subprim-address))

  instruction and makes the ba/bla branch to that instruction.
- this is the only use of the "constant pool" and there are no longer
  user-visible directives for referencing pc-relative data.  (We can
  load 32-bit integer constants via movw/movt instructions and initialize
  FPRs to constants via GPRs.)
- by default, the disassembler hides this and shows ba/bla instructions.

Compared to the scheme of a few days ago, it's about the same speed
(b/bl to LDR vs mov reg/bx reg).  If a subprim's called once per function
it's a little bigger; if there's more than one call site, it can be smaller.
(And we don't have to find a temp register.)  If we can map the subprims
to addresses within 32MB of the pure area, then purify can turn the PC-relative
branches/bls to the LDR instructions into direct branches/bls to the code.

Compared to the original scheme (branch/bl to mov pc, #n) we don't flush
the pipeline on every call and don't have any constraints on subprimitive
addresses (they don't have to be expressible as ARM constants.)