Try to allow some cases of self calls to skip arg-count checks, self-tail-calls to reuse existing frame.
When trying to get one or two forms into arbitrary registers,
try to better exploit cases where the forms are references to
variables that're already known to be in registers.

Work-in-progress: implement some forms of CASE with a jump table.
(Pretty simple on ARM; code doesn't yet deal with some contexts
- such as some forms of multiple-value return - and needs to
track stack changes, etc.

When trying to get the result of (%i+ val constant) into a register,
"suggest" that VAL be targeted to that register.  (This is just done
to avoid invalidating register map entries in some cases; sometimes
it wins, sometimes it doesn't.)