Back out of r14899 (ARM backend changes): at least one of those changes is bogus.  Some of those changes would be worthwhile if they were correct; I'll try to isolate and fix the problems.
Although the commit message neglected to mention it, in r14897 the assembler
stopped allowing arbitrary symbols when it really wanted to use keywords,
on the assumption that "we always used keywords anyhow".  That wasn't quite
true; change arm-vinsns, arm-lapmacros, and a few level-0 sources to ensure
that it is.  (I don't think that we were keywordizing at runtime when
compiling lisp code, so the change had minimal effect on anything but
arguable aesthetics.)