Windows version of RUN-PROGRAM and related functions: RUN-PROGRAM allows its "args" argument to be a SIMPLE-STRING (as well as a list of SIMPLE-STRINGs).
Use MAKE-WINDOWS-COMMAND-LINE (not just JOIN-STRINGS) to build the
command-line for CreateProcess from a cons of the RUN-PROGRAM "program
name" and "args" arguments.  A literal string "args" argument is passed
verbatim to the command line; the program name and any arguments in a
list-typed "args" arguments are processd by surrounding space/tab with
double-quotes and prefixing literal double-quote characters with backslash;
all other characters are passed verbatim to the command line and arguments
are separated by spaces.

This seems to be at least a first approximation of the rules used by
the MSVC runtime, though I haven't seen those rules written down
anywhere.  (The program name is processed by the OS and any quoting of
it has to be done in a canonical way.)

This may fix ticket:858.  How would anyone know ?