Sequences of one or more backslashes that precede a double-quote character or whitespace need to be escaped (doubled).
Fixes ticket:858 in the trunk.

Changes made to the trunk are often incorporated in the next release.
We generally assume that people can call REBUILD-CCL for themselves
when there aren't any bootstrapping issues involved (and there aren't
here); we do often provide kernel binaries for Windows since (a) it
can be tricky to install the necessary toolchain and support and (b)
the available Windows toolchains are often broken.