Try to clean up code which uses CMPXCHG: imm0 should contain the expected value (which may or may not be the current value, as of a few cycles before the CMPXCHG ...).  In general, we don't need or want to repeat the CMPXCHG in order to do a conditional store (failures aren't transient).  In cases where we repeat a CMPXCHG in a loop, ensure that the loop contains a PAUSE instruction to work correctly with hyperthreading.
Change the x86 pc_luser_xp() to account for changes in
_SPstore_node_conditional and _SPset_hash_key_conditional.

Introduce a WITH-EXCEPTION-LOCK macro; refactor
%LOCK-RECURSIVE-LOCK-OBJECT and friends so that we can lock/unlock a
kernel lock (with no lisp LOCK object around it) without having to
call into the kernel.  RECURSIVE-LOCK-WHOSTATE allows its argument to
be a string.  (WITH-EXCEPTION-LOCK isn't used anywhere yet; it may be
a better alternative to things like WITHOUT-GCING, where (a) it's
preferable to delay exception handing in other threads than to let
the heap grow and (b) the body is short and doesn't try to grab other
locks.)

This is all intended to fix ticket:1030 in the trunk.