Use an ACODE istruct to represent an acode node; such an istruct has OPERATOR, OPERANDS, ASSERTED-TYPE, and INFO fields.  (This allows us to avoid using a hash table to map between ACODE forms and code-/source-notes, and that in turn makes it easier to do certain transforms and optimizations on acode.
For this to be viable, we have to be much more rigorous than we
have been about using accessors (ACODE-OPERATOR, etc.) instead
of [%]CAR/[%]CDR chains and about using MAKE-ACODE rather than
CONS/LIST.  Catch enough of this in the frontend and X86 backend
so that the compiler can bootstrap itself on X86.

The acode accessor macros expand into calls to functions that
do some extra typechecking; this catches some errors at this
point but the plan is to remove these training wheels in the
near future.