Provide a specialized STREAM-READ-CHAR-NO-HANG on STRING-INPUT-STREAM. (The inherited method accesses the stream's ioblock.inbuf, which is null on STRING-INPUT-STREAMs.  That causes random bits to be accessed, and those random bits aren't really fixnums.)
Propagate the recent STRING-EQUAL fix from the trunk (so that we pass
the test suite.)

More compiler fixes, most notably in X862-FORM: we need to do things
that have side-effects even if result is unused.  (Need to do this
on other platforms as well.)  This was masked by the fact that many
acode operators that could/should have had their acode-subforms bit
set didn't.