Change the type system's assumptions about the ordering of typecodes/subtags that represent CL vectors and arrays.  We now assume:
a) for gvectors: target::subtag-simple-vector > target::subtag-vectorH >
   target::subtag-arrayH > all other gvector typecodes
b) for ivectors: all ivector subtags that denote CL arrays are >=
   target::min-cl-ivector-subtag and all other ivector subtags are
   less than that value
c) there is no defined ordering between ivector subtags that denote
   CL vectors and gvector subtypes that do so; target::min-array-subtag
   and target::min-vector-subtag no longer exist.

In general, this slightly complicates things like ARRAYP and VECTORP but
lets us define several more ivector and gvector types, which have been
in especially short supply on 32-bit platforms.

This is all implemented on the ARM, and seems to work reasonably.

Various forms of ARRAYP and VECTORP now have to treat the ivector
and gvector cases seperately; this is currently implemented by use
of new primitives IVECTOR-TYPECODE-P and GVECTOR-TYPECODE-P, which
return their argument if it's an ivector (gvector) typecode and 0
otherwise.  (These operations will be a little more complicated on
64-bit platforms, where we have multiple ivector/gvector fulltags.)