Don't use SPILL-AND-SPLIT-INTERVAL to deal with intervals whose physical registers are killed by a CALL; simply treat those registers as caller-save. I am not certain why, but cases like
(if something (call) no-call) 

didn't work correctly if the spill-and-split happened on one arm of the
branch and not on the other.

a CALL is just something that increases register pressure, so we should
check to ensure that the same problem doesn't occur when that pressure is
dealt with in the same way when the pressure arises in other ways.