Add :both case to #'shutdown.
Redefine the class called "unconnected-sockets" to be "device-mixin."
First because it was creating unnecessary diamonds in the class
lattice, and second because connected UDP sockets do in fact
exist. They are defined as UDP sockets on which #_connect has been
called. And if you don't keep track of which UDP sockets are connected
in this sense, calling #_sendto on them while supplying a non-null
address will throw an EISCONN (56) error on OSX, even though it's usually
tolerated on Linux.

Many sites on the web confirm this. The manpage for sendto is
misleading when it states that sendto() may be used at any time. It
*can* in fact be used at any time, but when used on a connected UDP
socket its fifth argument MUST be a null pointer, as stated in
Stevens' "UNIX Network Programming V1", Section 8.11, and, according to
Stevens, the Posix spec.

Fixes Ticket #1252.