Define FIXED-LREG-P, which retrns true if an lreg is wired or vinsn-local; use it to avoid a bug in RESOLVE-NON-YTIVIAL-INTERVAL-CONFLICTS.
try to avoid nullifying a trivial-copy instruction in cases
where the destination logical register is defined more than once and referenced
after a subsequent definition when neither the redefinition or reference is
itseld a trivial-copy.   This may or may not be overly conservative, but avoids 
some bugs when it is not: see STRING-INPUT-STREAM-IOBLOCK-READ-CHAR, which otherwise
can effectively change the value of IDX incorrectly.